Fix floating point rounding

+++ cache/hhvm-scalable-cache.h
@@ -204,7 +204,7 @@ private:
   std::queue<int> construct_container; // Note that this is a queue not a stack
   int PASS_THRESHOLD = 3;
 
-  int MELT_CHUNK_FRACTION = 0.2;
+  float MELT_CHUNK_FRACTION = 0.2;
   int COUNT_THRESHOLD = 2;
   double performance_depletion = COUNT_THRESHOLD;
   double best_sleep = 0;
